+++ R/convert_d2logit.R
@@ -1,27 +1,15 @@
-#' @title Convert effect size d into OR
-#' @name esc_d2or
+#' @title Convert effect size d into log odds
+#' @name convert_d2logit
 #'
-#' @description Compute effect size \code{OR} from effect size \code{d}.
+#' @description Compute effect size \code{log odds} from effect size \code{d}.
 #'
-#' @param d The effect size \code{d}.
-#' @param se The standard error of \code{d}. One of \code{se} or \code{v}
-#'        must be specified.
-#' @param v The variance of \code{d}. One of \code{se} or \code{v} must be
-#'        specified.
-#' @param es.type Type of effect size odds ratio that should be returned.
-#'        May be \code{es.type = "logit"} or \code{es.type = "cox"}
-#'        (see 'Details').
-#' @param info String with information on the transformation. Used for the
-#'        print-method. Usually, this argument can be ignored
-#'
+#' @inheritParams convert_d2or
 #' @inheritParams esc_beta
 #' @inheritParams hedges_g
 #'
-#' @note Effect size is returned as \code{exp(log_values)} (odds ratio),
-#'       confidence intervals are also exponentiated. To get the log-values,
-#'       use \code{\link{esc_d2logit}}.
-#'       \strong{However}, variance and standard error of this function
-#'       are returned on the log-scale!
+#' @note Effect size, variance, standard error and confidence intervals are
+#'       returned on the log-scale. To get the odds ratios and exponentiated
+#'       confidence intervals, use \code{\link{convert_d2or}}.
 #'
 #' @details Conversion from \code{d} to odds ratios can be done with two
 #'          methods:
@@ -44,19 +32,19 @@
 #'             Cox DR. 1970. Analysis of binary data. New York: Chapman & Hall/CRC
 #'             \cr \cr
 #'             Hasselblad V, Hedges LV. 1995. Meta-analysis of screening and diagnostic tests. Psychological Bulletin 117(1): 167–178. \doi{10.1037/0033-2909.117.1.167}
-#'             \cr \cr
-#'             Borenstein M, Hedges LV, Higgins JPT, Rothstein HR. 2009. Introduction to Meta-Analysis. Chichester, West Sussex, UK: Wiley
 #'
 #' @examples
-#' # d to odds ratio
-#' esc_d2or(0.7, se = 0.5)
-#' # odds ratio to d
-#' esc_or2d(3.56, se = 0.91)
+#' # to logits
+#' convert_d2logit(0.7, se = 0.5)
+#'
+#' # to Cox-logits
+#' convert_d2logit(0.7, v = 0.25, es.type = "cox")
 #'
 #' @export
-esc_d2or <- function(d, se, v, totaln,
-                     es.type = c("logit", "cox"),
-                     info = NULL, study = NULL) {
+convert_d2logit <- function(d, se, v, totaln,
+                        es.type = c("logit", "cox"),
+                        info = NULL,
+                        study = NULL) {
   # match  arguments
   es.type <- match.arg(es.type)
 
@@ -67,44 +55,34 @@ esc_d2or <- function(d, se, v, totaln,
   }
 
   # do we have se?
-  if (!missing(se) && !is.null(se) && !is.na(se)) v <- se ^ 2
+  if (!missing(se) && !is.null(se) && !is.na(se)) v <- se^2
 
   # do we have total n?
   if (missing(totaln) || is.na(totaln)) totaln <- NULL
 
   # do we have a separate info string?
   if (is.null(info)) {
-    info <- "effect size d to effect size OR"
+    info <- "effect size d to effect size logit"
     if (es.type == "cox") info <- paste0(info, "(Cox)")
   }
 
   if (es.type == "logit") {
     # Hasselblad and Hedges logit
     es <- pi / sqrt(3) * d
-    v <- (pi  ^ 2) / 3 * v
-    measure <- "or"
+    v <- (pi^2) / 3 * v
+    measure <- "logit"
   } else {
     # Cox logit
     es <- d * 1.65
     v <- v / .367
-    measure <- "cox-or"
+    measure <- "cox-logit"
   }
 
   # return effect size d
   structure(
-    class = c("esc", "esc_d2or"),
-    list(
-      es = exp(es),
-      se = sqrt(v),
-      var = v,
-      ci.lo = exp(lower_d(es, v)),
-      ci.hi = exp(upper_d(es, v)),
-      w = 1 / v,
-      totaln = totaln,
-      measure = measure,
-      info = info,
-      study = study
-    )
+    class = c("esc", "convert_d2logit"),
+    list(es = es, se = sqrt(v), var = v, ci.lo = lower_d(es, v),
+         ci.hi = upper_d(es, v), w = 1 / v, totaln = totaln,
+         measure = measure, info = info, study = study)
   )
 }
-