+++ R/convert_z2r.R
@@ -1,5 +1,5 @@
 #' @title Convert correlation coefficient r into Fisher's z
-#' @name esc_r2z
+#' @name convert_r2z
 #'
 #' @description Convert correlation coefficient r into Fisher's z.
 #'
@@ -12,14 +12,14 @@
 #'             Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University
 #'
 #' @examples
-#' esc_r2z(.03)
+#' convert_r2z(.03)
 #'
 #' @export
-esc_r2z <- function(r) return(.5 * log((1 + r) / (1 - r)))
+convert_r2z <- function(r) return(.5 * log((1 + r) / (1 - r)))
 
 
 #' @title Convert Fisher's z into correlation coefficient r
-#' @name esc_z2r
+#' @name convert_z2r
 #'
 #' @description Convert Fisher's z into correlation coefficient r.
 #'
@@ -32,10 +32,10 @@ esc_r2z <- function(r) return(.5 * log((1 + r) / (1 - r)))
 #'             Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University
 #'
 #' @examples
-#' esc_z2r(.03)
+#' convert_z2r(.03)
 #'
 #' @export
-esc_z2r <- function(z) return((exp(2 * z) - 1) / (exp(2 * z) + 1))
+convert_z2r <- function(z) return((exp(2 * z) - 1) / (exp(2 * z) + 1))